Modernize Autotools usage, real instructions
The code has bitrotted somewhat since the last commit, and
frankly, Autotools has never been a friendly technology.
But this should clean up the way it handles on modern
systems, and give users a bit more guidance in setting up
this synth.

That's not to say it will be a guaranteed smooth ride, but
it never hurts to do a bit of dusting and sand the
sharpness off some edges.